+++ src/domain/tab/bottom/components/BottomTabs.tsx
@@ -9,7 +9,7 @@ import {
 } from "../types/bottomTabActiveType";
 import TabItem from "./BottomTabItem";
 import { backgroundColor } from "../../../../common/styles/color";
-import { useBottomTabActive } from "../hooks/bottomTabActiveHooks";
+import { useBottomTabActive } from "../hooks/useBootomTabActive";
 
 export default function BottomTabs() {
   const { bottomTabActive, setTabActive } = useBottomTabActive();
